chore: split out compressor model types
