+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/consumer_function_mapper/compressor_consumer_function.py
@@ -2,7 +2,7 @@ from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_f
     CompressorConsumerFunction,
 )
 from libecalc.domain.process.compressor.core import create_compressor_model
-from libecalc.domain.process.compressor.dto.base import CompressorConsumerFunction as CompressorConsumerFunctionDTO
+from libecalc.domain.process.compressor.dto import CompressorConsumerFunction as CompressorConsumerFunctionDTO
 
 
 def create_compressor_consumer_function(model_dto: CompressorConsumerFunctionDTO) -> CompressorConsumerFunction:
