+++ src/libecalc/domain/process/compressor/core/train/single_speed_compressor_train_common_shaft.py
@@ -19,7 +19,7 @@ from libecalc.domain.process.compressor.core.train.utils.numeric_methods import 
     find_root,
     maximize_x_given_boolean_condition_function,
 )
-from libecalc.domain.process.compressor.dto.train import SingleSpeedCompressorTrain
+from libecalc.domain.process.compressor.dto import SingleSpeedCompressorTrain
 from libecalc.domain.process.core.chart.chart_area_flag import ChartAreaFlag
 from libecalc.domain.process.core.results.compressor import TargetPressureStatus
