+++ src/libecalc/presentation/yaml/mappers/consumer_function_mapper.py
@@ -5,11 +5,11 @@ from libecalc.common.energy_usage_type import EnergyUsageType
 from libecalc.common.time_utils import Period, define_time_model_for_period
 from libecalc.common.utils.rates import RateType
 from libecalc.domain.process.compressor.dto import (
-    CompressorModelTypes,
+    CompressorConsumerFunction,
+    CompressorWithTurbine,
     VariableSpeedCompressorTrainMultipleStreamsAndPressures,
 )
-from libecalc.domain.process.compressor.dto.base import CompressorConsumerFunction
-from libecalc.domain.process.compressor.dto.with_turbine import CompressorWithTurbine
+from libecalc.domain.process.compressor.dto.model_types import CompressorModelTypes
 from libecalc.domain.process.dto import (
     ConsumerFunction,
     DirectConsumerFunction,
